Fix per-frame allocations and FormatDuration overflow
TimelineBuilder: build vesselNameById dictionary once in Build() and
pass to both collectors instead of building it twice. Replace LINQ
OrderBy().ToList() with in-place List.Sort. Remove System.Linq using.

TimelineEntryDisplay: change FormatDuration locals from int to long to
prevent overflow for durations exceeding ~68 years (int max ~2.1B sec).

+++ Source/Parsek/Timeline/TimelineEntryDisplay.cs
@@ -107,15 +107,15 @@ namespace Parsek
             if (seconds <= 0) return "";
 
             // KSP uses 6-hour days, 426-day years (Kerbin calendar)
-            int totalSeconds = (int)seconds;
-            int s = totalSeconds % 60;
-            int totalMinutes = totalSeconds / 60;
-            int m = totalMinutes % 60;
-            int totalHours = totalMinutes / 60;
-            int h = totalHours % 6;       // KSP day = 6 hours
-            int totalDays = totalHours / 6;
-            int d = totalDays % 426;       // KSP year = 426 days
-            int y = totalDays / 426;
+            long totalSeconds = (long)seconds;
+            long s = totalSeconds % 60;
+            long totalMinutes = totalSeconds / 60;
+            long m = totalMinutes % 60;
+            long totalHours = totalMinutes / 60;
+            long h = totalHours % 6;       // KSP day = 6 hours
+            long totalDays = totalHours / 6;
+            long d = totalDays % 426;       // KSP year = 426 days
+            long y = totalDays / 426;
 
             var parts = new System.Collections.Generic.List<string>(5);
             if (y > 0) parts.Add($"{y}y");
